jd_generator/requirements.txt
fastapi>=0.105.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.3.0 